Removed pilosa-dependent files from Windows compilation
